Use lookup table for flipped SNPs
I experimented with a number of different approaches, but this seemed the
best; however, in my simple microbenchmarking this approach is slower
than previous commit using a conditional and smaller _bytemap. I don't
understand how that can be correct, unless it is related to the cache in
some way. In any case, it is ~1 ns difference; it may be worth switching
back to the conditional approach.